# Faster quaternion multiplication