--- a/README.md
+++ b/README.md
@@ -54,6 +54,7 @@ Where each group supports:
   **`adjoint()`**, **`apply()`**, **`multiply()`**, **`inverse()`**,
   **`identity()`**, **`from_matrix()`**, and **`as_matrix()`** operations. (see
   [./examples/se3_example.py](./examples/se3_basics.py))
+- Taylor approximations near singularities.
 - Helpers for optimization on manifolds (see
   [./examples/se3_optimization.py](./examples/se3_optimization.py),
   <code>jaxlie.<strong>manifold.\*</strong></code>).
@@ -78,24 +79,10 @@ pip install jaxlie
 
 ---
 
-### In the wild
-
-- [jaxfg](https://github.com/brentyi/jaxfg) applies `jaxlie` to nonlinear least
-  squares problems with block-sparse structure. (for pose graph optimization,
-  bundle adjustment, etc)
-- [tensorf-jax](https://github.com/brentyi/tensorf-jax) is an unofficial
-  implementation of
-  [Tensorial Radiance Fields (Chen et al, ECCV 2022)](https://apchenstu.github.io/TensoRF/)
-  using `jaxlie`.
-  ![Render of a lego](https://github.com/brentyi/tensorf-jax/raw/main/lego_render.gif)
-
----
-
 ### Misc
 
-`jaxlie` was originally written for our IROS 2021 paper
-([link](https://github.com/brentyi/dfgo)). If it's useful for you, you're
-welcome to cite:
+`jaxlie` was originally written when I was learning about Lie groups for our IROS 2021 paper
+([link](https://github.com/brentyi/dfgo)):
 
 ```
 @inproceedings{yi2021iros,