--- a/tests/test_broadcast.py
+++ b/tests/test_broadcast.py
@@ -3,12 +3,7 @@
 from typing import Tuple, Type
 
 import numpy as onp
-from hypothesis import given, settings
-from hypothesis import strategies as st
-from jax import numpy as jnp
 from utils import (
-    assert_arrays_close,
-    assert_transforms_close,
     general_group_test,
     sample_transform,
 )