@brentyi 2024-11-28
Runtime changes in [EgoAllo](https://github.com/brentyi/egoallo):
- JIT: 90.1 seconds => 49.8 seconds
- Optimization: 0.178 => 0.155 seconds